+++ packages/api/src/trpc.ts
@@ -32,7 +32,9 @@ export const createTRPCContext = (opts: {
   const session = opts.session;
   const source = opts.headers.get("x-trpc-source") ?? "unknown";
 
-  console.log(">>> tRPC Request from", source, "by", session?.user);
+  console.log(
+    `>>> tRPC Request from ${source} by ${session ? `user id ${session.user.id}` : `anonymous user`}`,
+  );
 
   return {
     session,
