+++ packages/auth/src/index.ts
@@ -1,6 +1,7 @@
 import NextAuth from "next-auth";
 
-import { authConfig, createAuthWithDecryptedUser } from "./config";
+import { authConfig } from "./config";
+import { createAuthWithDecryptedUser } from "./utils";
 
 export type { Session } from "next-auth";
